Graphs showing the relation of shape and species
Figure i explains the scattered plot of the relation of the bird species with the pointiness,  which determines the deviation from the perfect spherical shape and the breadth of the egg

Figure ii explains the relation between eggs in a nest, birds height, and body mass. The more eggs in a nest more spherical the egg is for more efficient warmth conservation.

Figure iii explains the relation between pointiness, breadth, and body mass of the bird species. The more the body mass of the bird the higher is the pointiness showing the more efficient use of space for higher grith.